+++ src/main/java/com/ae2powertools/ItemRegistry.java
@@ -12,6 +12,7 @@ import net.minecraftforge.fml.relauncher.SideOnly;
 
 import com.ae2powertools.items.ItemCardsDistributor;
 import com.ae2powertools.items.ItemCrafterSpeedUpgrade;
+import com.ae2powertools.items.ItemLevelMonitorAlarmLocator;
 import com.ae2powertools.items.ItemNetworkComponentLocator;
 import com.ae2powertools.items.ItemNetworkHealthScanner;
 import com.ae2powertools.items.ItemPriorityTuner;
@@ -33,6 +34,7 @@ public class ItemRegistry {
     public static ItemPriorityTuner PRIORITY_TUNER;
     public static ItemCardsDistributor CARDS_DISTRIBUTOR;
     public static ItemRemoteStorageMonitor REMOTE_STORAGE_MONITOR;
+    public static ItemLevelMonitorAlarmLocator LEVEL_MONITOR_ALARM_LOCATOR;
     public static ItemCrafterSpeedUpgrade CRAFTER_SPEED_UPGRADE;
     public static ItemPartStorageLevelEmitter STORAGE_LEVEL_EMITTER_PART;
     public static ItemPartStorageDisplay STORAGE_DISPLAY_PART;
@@ -45,6 +47,7 @@ public class ItemRegistry {
         PRIORITY_TUNER = new ItemPriorityTuner();
         CARDS_DISTRIBUTOR = new ItemCardsDistributor();
         REMOTE_STORAGE_MONITOR = new ItemRemoteStorageMonitor();
+        LEVEL_MONITOR_ALARM_LOCATOR = new ItemLevelMonitorAlarmLocator();
         CRAFTER_SPEED_UPGRADE = new ItemCrafterSpeedUpgrade();
         STORAGE_LEVEL_EMITTER_PART = new ItemPartStorageLevelEmitter();
         STORAGE_DISPLAY_PART = new ItemPartStorageDisplay();
@@ -60,6 +63,7 @@ public class ItemRegistry {
             PRIORITY_TUNER,
             CARDS_DISTRIBUTOR,
             REMOTE_STORAGE_MONITOR,
+            LEVEL_MONITOR_ALARM_LOCATOR,
             CRAFTER_SPEED_UPGRADE,
             STORAGE_LEVEL_EMITTER_PART,
             STORAGE_DISPLAY_PART,
@@ -76,6 +80,7 @@ public class ItemRegistry {
         registerItemModel(PRIORITY_TUNER);
         registerItemModel(CARDS_DISTRIBUTOR);
         registerItemModel(REMOTE_STORAGE_MONITOR);
+        registerItemModel(LEVEL_MONITOR_ALARM_LOCATOR);
         registerItemModel(STORAGE_LEVEL_EMITTER_PART);
         registerItemModel(STORAGE_DISPLAY_PART);
         registerItemModel(STORAGE_DISPLAY_SMALLER_PART);